refactor(pdf): extract note preclassification
